app_web/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
minio==7.2.3
Flask-CORS==4.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.3
MarkupSafe==2.1.5
urllib3==2.2.1
certifi==2024.2.2
chardet==5.2.0
idna==3.7
requests==2.31.0
PyJWT==2.8.0 